requirements.txt
fastapi>=0.116.1
python-dotenv>=1.1.1
python-multipart>=0.0.20
requests>=2.31.0
uvicorn>=0.35.0
yt-dlp>=2025.10.22
whisperx>=3.1.1
apscheduler>=3.10.4